+++ mobile/src/url/variable.js
@@ -1,9 +1,10 @@
 import axios from "axios";
 const carmela = "http://192.168.0.73:3001/api";
+const gerar = "http://192.168.1.85:3001/api"
 
 
 export const URLBase = axios.create({
-    baseURL: carmela
+    baseURL: gerar
 })
